modulo rol: GET, POST completado

+++ BiblioCleanSol/BiblioCleanSol.Application/Dtos/Usuarios/Rol/RolEditarDto.cs
@@ -2,7 +2,7 @@
 {
     public record RolEditarDto : RolDto
     {
-        public int RolId { get; set; }
+        public byte RolId { get; set; }
         //Campos de auditoria
         public DateTime FechaMod { get; set; }
         public int UsuarioModId { get; set; }
